Update to 13.20

+++ t/Writer.t
@@ -1114,20 +1114,30 @@ my $testOK;
 # test 61: Test the %f and %z date/time format codes for writing
 {
     ++$testnum;
-    my $exifTool = Image::ExifTool->new;
-    $exifTool->Options(DateFormat => '%H:%M:%S%f%:z %m-%d-%Y');
-    $exifTool->SetNewValue('XMP:CreateDate' => '09:11:12.345-07:00 01-02-2024');
-    $testfile = "t/${testname}_${testnum}_failed.xmp";
-    unlink $testfile;
-    writeInfo($exifTool, undef, $testfile);
-    $exifTool->Options(DateFormat => undef);
-    my $info = $exifTool->ImageInfo($testfile, '-system:all');
-    if (check($exifTool, $info, $testname, $testnum)) {
+    my $skip = '';
+    if (eval { require POSIX::strptime } or eval { require Time::Piece }) {
+        my $exifTool = Image::ExifTool->new;
+        $exifTool->Options(DateFormat => '%H:%M:%S%f%:z %m-%d-%Y');
+        $exifTool->SetNewValue('XMP:CreateDate' => '09:11:12.345-07:00 01-02-2024');
+        $testfile = "t/${testname}_${testnum}_failed.xmp";
         unlink $testfile;
+        my $success = writeInfo($exifTool, undef, $testfile);
+        if ($success) {
+            $exifTool->Options(DateFormat => undef);
+            my $info = $exifTool->ImageInfo($testfile, '-system:all');
+            $success = check($exifTool, $info, $testname, $testnum);
+        }
+        if ($success) {
+            unlink $testfile;
+        } else {
+            notOK();
+        }
     } else {
-        notOK();
+        warn "\n  POSIX::strptime or Time::Piece not installed, so you won't\n";
+        warn "  have the abilty use the DateFormat feature when writing.\n";
+        $skip = ' # POSIX::strptime or Time::Piece not installed';
     }
-    print "ok $testnum\n";
+    print "ok $testnum$skip\n";
 }
 
 done(); # end
